refactor into constants/metadata, add parse_href function
