+++ pyproject.toml
@@ -22,6 +22,7 @@ requires-python = ">=3.8"
 dependencies = [
   "httpx~=0.27.0",
   "pandas~=2.2.2",
+  "parse~1.20.2",
   "pystac~=1.10.1",
   "rasterio~=1.3.10",
   "stactools>=0.4.0",
