+++ src/stactools/noaa_hrrr/commands.py
@@ -5,8 +5,8 @@ from datetime import datetime
 import click
 from click import Command, Group
 from stactools.noaa_hrrr import stac
-from stactools.noaa_hrrr.constants import (
-    EXTENDED_FORECAST_MAX_HOUR,
+from stactools.noaa_hrrr.constants import EXTENDED_FORECAST_MAX_HOUR
+from stactools.noaa_hrrr.metadata import (
     CloudProvider,
     Product,
     Region,
