+++ tests/test_commands.py
@@ -5,7 +5,7 @@ from click import Group
 from click.testing import CliRunner
 from pystac import Collection, Item
 from stactools.noaa_hrrr.commands import create_noaahrrr_command
-from stactools.noaa_hrrr.constants import (
+from stactools.noaa_hrrr.metadata import (
     CloudProvider,
     Product,
     Region,
